refactor: Merge `Handler` and `Service`.
