+++ spec/EcPhp/CasLib/Introspection/IntrospectorSpec.php
@@ -219,6 +219,55 @@ class IntrospectorSpec extends ObjectBehavior
             ->during('parse', [$response, 'FOOBAR']);
     }
 
+    public function it_can_detect_the_type_of_a_response()
+    {
+        $body = <<< 'EOF'
+            <cas:serviceResponse xmlns:cas="http://www.yale.edu/tp/cas">
+             <cas:authenticationSuccess>
+              <cas:user>username</cas:user>
+             </cas:authenticationSuccess>
+            </cas:serviceResponse>
+            EOF;
+
+        $headers = [
+            'Content-Type' => 'application/xml',
+        ];
+
+        $response = new Response(
+            200,
+            $headers,
+            $body
+        );
+
+        $this
+            ->detect(
+                $response
+            )
+            ->shouldReturnAnInstanceOf(ServiceValidate::class);
+
+        $body = <<< 'EOF'
+            <cas:serviceResponse xmlns:cas="http://www.yale.edu/tp/cas">
+             <cas:authenticationSuccess>
+              <cas:user>username</cas:user>
+             </cas:authenticationSuccess>
+            </cas:serviceResponse>
+            EOF;
+
+        $headers = [
+            'Content-Type' => 'application/foo',
+        ];
+
+        $response = new Response(
+            200,
+            $headers,
+            $body
+        );
+
+        $this
+            ->shouldThrow(Exception::class)
+            ->during('detect', [$response]);
+    }
+
     public function it_is_initializable()
     {
         $this->shouldHaveType(Introspector::class);
