+++ src/Handler/Handler.php
@@ -12,17 +12,27 @@ declare(strict_types=1);
 namespace EcPhp\CasLib\Handler;
 
 use EcPhp\CasLib\Configuration\PropertiesInterface;
+use EcPhp\CasLib\Introspection\Contract\IntrospectorInterface;
 use EcPhp\CasLib\Utils\Uri;
+use Exception;
 use loophp\psr17\Psr17Interface;
 use Psr\Cache\CacheItemPoolInterface;
+use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\UriInterface;
 
 use function array_key_exists;
 
+use const JSON_ERROR_NONE;
+
 abstract class Handler
 {
     private CacheItemPoolInterface $cache;
+
+    private ClientInterface $client;
+
+    private IntrospectorInterface $introspector;
 
     private array $parameters;
 
@@ -35,11 +45,15 @@ abstract class Handler
      */
     public function __construct(
         array $parameters,
+        CacheItemPoolInterface $cache,
+        ClientInterface $client,
+        IntrospectorInterface $introspector,
         PropertiesInterface $properties,
-        Psr17Interface $psr17,
-        CacheItemPoolInterface $cache
+        Psr17Interface $psr17
     ) {
         $this->cache = $cache;
+        $this->client = $client;
+        $this->introspector = $introspector;
         $this->parameters = $parameters;
         $this->properties = $properties;
         $this->psr17 = $psr17;
@@ -77,13 +91,7 @@ abstract class Handler
         // Filter out empty $query parameters
         $query = array_filter(
             $query,
-            static function ($item): bool {
-                if ([] === $item) {
-                    return false;
-                }
-
-                return '' !== $item;
-            }
+            static fn ($item): bool => [] === $item ? false : ('' !== $item)
         );
 
         return $this->getPsr17()
@@ -100,15 +108,9 @@ abstract class Handler
      */
     protected function formatProtocolParameters(array $parameters): array
     {
-        $parameters = array_filter(
-            $parameters
-        );
-
         $parameters = array_map(
-            static function ($parameter) {
-                return true === $parameter ? 'true' : $parameter;
-            },
-            $parameters
+            static fn ($parameter) => true === $parameter ? 'true' : $parameter,
+            array_filter($parameters)
         );
 
         if (true === array_key_exists('service', $parameters)) {
@@ -130,6 +132,16 @@ abstract class Handler
     protected function getCache(): CacheItemPoolInterface
     {
         return $this->cache;
+    }
+
+    protected function getClient(): ClientInterface
+    {
+        return $this->client;
+    }
+
+    protected function getIntrospector(): IntrospectorInterface
+    {
+        return $this->introspector;
     }
 
     /**
@@ -158,5 +170,43 @@ abstract class Handler
     protected function getPsr17(): Psr17Interface
     {
         return $this->psr17;
+    }
+
+    /**
+     * Normalize a response.
+     */
+    protected function normalize(RequestInterface $request, ResponseInterface $response): ResponseInterface
+    {
+        $body = $this->parse($request, $response);
+
+        if ([] === $body) {
+            throw new Exception('Unable to parse the response during the normalization process.');
+        }
+
+        $body = json_encode($body);
+
+        if (false === $body || JSON_ERROR_NONE !== json_last_error()) {
+            throw new Exception('Unable to encode the response in JSON during the normalization process.');
+        }
+
+        return $response
+            ->withBody($this->getPsr17()->createStream($body))
+            ->withHeader('Content-Type', 'application/json');
+    }
+
+    /**
+     * Parse the response format.
+     *
+     * @return array[]|string[]
+     *   The parsed response.
+     */
+    protected function parse(RequestInterface $request, ResponseInterface $response): array
+    {
+        return $this
+            ->getIntrospector()
+            ->parse(
+                $response,
+                $this->getProtocolProperties($request)['default_parameters']['format'] ?? 'XML'
+            );
     }
 }
