+++ src/Handler/ServiceValidateHandlerInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ *
+ * @see https://github.com/ecphp
+ */
+
+declare(strict_types=1);
+
+namespace EcPhp\CasLib\Handler;
+
+interface ServiceValidateHandlerInterface extends HandlerInterface
+{
+}
