+++ src/Introspection/Introspection.php
@@ -11,7 +11,6 @@ declare(strict_types=1);
 
 namespace EcPhp\CasLib\Introspection;
 
-use EcPhp\CasLib\Introspection\Contract\IntrospectionInterface;
 use Psr\Http\Message\ResponseInterface;
 
 abstract class Introspection
@@ -42,13 +41,5 @@ abstract class Introspection
     public function getResponse(): ResponseInterface
     {
         return $this->response;
-    }
-
-    public function withParsedResponse(array $parsedResponse): IntrospectionInterface
-    {
-        $clone = clone $this;
-        $clone->parsedResponse = $parsedResponse;
-
-        return $clone;
     }
 }
