+++ tests/Cas.php
@@ -12,7 +12,6 @@ declare(strict_types=1);
 namespace tests\EcPhp\CasLib;
 
 use EcPhp\CasLib\CasInterface;
-use EcPhp\CasLib\Configuration\PropertiesInterface;
 use EcPhp\CasLib\Introspection\Contract\IntrospectionInterface;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
@@ -36,11 +35,6 @@ final class Cas implements CasInterface
         ResponseInterface $response
     ): IntrospectionInterface {
         return $this->cas->detect($response);
-    }
-
-    public function getProperties(): PropertiesInterface
-    {
-        return $this->cas->getProperties();
     }
 
     public function handleProxyCallback(
